카드 리스트 출력

+++ src/main/java/upload/util/DirectoryManagement.java
@@ -111,10 +111,13 @@ public class DirectoryManagement {
 	public void deleteFile(HttpServletRequest request, int no) {
 		String path = request.getSession().getServletContext().getRealPath("/card/temp/"+no);
 		File file = new File(path);
-		File[] fileList = file.listFiles();
-		if(fileList.length != 0) {
-			for(File f : fileList) {
-				f.delete();
+		if(file.exists()) {
+			File[] fileList = file.listFiles();
+			System.out.println(fileList.length);
+			if(fileList.length != 0) {
+				for(File f : fileList) {
+					f.delete();
+				}
 			}
 		}
 	}
